♻️ Q8W0E9 pr: decompose sync batch ownership
