+++ .agentplane/tasks/202605290353-Q8W0E9/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290353-Q8W0E9`
+Title: PR sync batch ownership decomposition
+Canonical task record: `.agentplane/tasks/202605290353-Q8W0E9/README.md`
+
+## Summary
+
+PR sync batch ownership decomposition
+
+Decompose packages/agentplane/src/commands/pr/internal/sync.ts by extracting branch_pr batch ownership helpers into a focused internal module, preserving PR sync behavior while reducing runtime hotspot count.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/commands/pr/internal/sync.ts by extracting branch_pr batch ownership helpers into a focused internal module, preserving PR sync behavior while reducing runtime hotspot count.
+- Out of scope: unrelated refactors not required for "PR sync batch ownership decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T03:53:37.038Z
+- Branch: task/202605290353-Q8W0E9/pr-sync-batch-ownership-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
