+++ .agentplane/tasks/202605290353-Q8W0E9/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 Q8W0E9 task: PR sync batch ownership decomposition [202605290353-Q8W0E9]
