+++ .agentplane/tasks/202605290353-Q8W0E9/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-29T03:53:37.038Z
+
+## Task
+
+- Task: `202605290353-Q8W0E9`
+- Title: PR sync batch ownership decomposition
+- Status: DOING
+- Branch: `task/202605290353-Q8W0E9/pr-sync-batch-ownership-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290353-Q8W0E9/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T03:53:37.038Z
+- Branch: task/202605290353-Q8W0E9/pr-sync-batch-ownership-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
